[fix](be) Avoid including multi-string search source in test
### What problem does this PR solve?

Issue Number: None

Related PR: None

Problem Summary: The multi-string search unit test included the production .cpp file directly. This violates build hygiene and can cause duplicate strong symbols in unity builds. Move the testable implementation template into a header shared by production code and the unit test.

### Release note

None

### Check List (For Author)

- Test: Unit Test
    - ./run-be-ut.sh --run --filter=FunctionsMultiStringSearchTest.* -j50 --coverage
    - ./build.sh --be
    - build-support/check-format.sh
    - build-support/run-clang-tidy.sh for the modified production and test translation units
- Behavior changed: No
- Does this need documentation: No
